refactor: toEntity 메서드 생성
- User, Article toEntity 메서드 생성

+++ src/main/java/com/kakao/cafe/article/repository/ArticleRepository.java
@@ -16,8 +16,8 @@ public class ArticleRepository {
         articles.add(article);
     }
 
-    public int getNumberOfArticles(){
-        return articles.size();
+    public Long getNumberOfArticles(){
+        return (long) articles.size();
     }
 
     public List<Article> findAll(){
